re-structured to match example, deleted react files
