+++ app/src/index.js
@@ -1,9 +1,19 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
+const express = require('express');
+const app = express();
 
-ReactDOM.hydrate(<App/>, document.getElementById("root"));
+// Serve static files
+ //app.use('/styles', express.static('/public/App.css'));
+// app.use('/scripts', express.static('path-to-your-js-folder'));
 
-serviceWorker.unregister();
+// set different properties 
+app.set('view engine', 'ejs');
+app.set('views', './src/views');
+
+// Render EJS template
+app.get('/register', (req, res) => {
+    res.render('registerView');
+});
+
+// Start the server
+const PORT = process.env.PORT || 3005;
+app.listen(PORT, () => console.log(`Server is running on port ${PORT}`));
