--- app/ssr/index.js
@@ -1,8 +1,0 @@
-require('ignore-styles')
-
-require('@babel/register')({
-    ignore: [/(node_modules)/],
-    presets: ['@babel/preset-env', '@babel/preset-react']
-})
-
-require('./server')
