+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "ejs": "^3.1.9"
+  }
+}
